Move review surface contracts
